+++ Project_Work_MAUI/Models/TransactionRequest.cs
@@ -3,7 +3,7 @@
     public class TransactionRequest
     {
         public string? iban { get; set; }
-        public int amount { get; set; }
+        public float amount { get; set; }
         public string categoryid { get; set; }
         public string description { get; set; }
     }
